refactor(console): merge utils/ into lib/
The three utils/ modules (bytes, loadChart, planLabel) are the same kind of pure
helper as everything already in lib/. Fold them in and drop the directory.
