+++ dashboard/src/lib/loadChart.ts
